refactor: align variable names as per intent
