+++ kotlin/openID4VP/src/commonMain/kotlin/io/mosip/openID4VP/authorizationResponse/unsignedVPToken/types/mdoc/UnsignedMdocVPTokenBuilder.kt
@@ -99,8 +99,6 @@ internal class UnsignedMdocVPTokenBuilder(
         return Pair(docTypeToDeviceAuthenticationBytes, unsignedVPTokens)
     }
 
-    // TODO: Add method func build(credentialToCredentialQueryIdMappings: inout [CredentialToCredentialQueryIdMapping])
-
     private sealed class MdocSpecVersionHandler {
         object Draft23 : MdocSpecVersionHandler()
         object SpecV1 : MdocSpecVersionHandler()
